- fixed migration scripts

+++ database/migrations/2022_10_29_140907_add_columns_to_check_payments_table.php
@@ -19,8 +19,11 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('check_payments', function (Blueprint $table) {
-            $table->dropColumn('status');
-            $table->dropColumn('bounced_deposit_id');
+            if (Schema::hasColumn('check_payments','status') === true) {
+                $table->dropColumn('status');
+            }
+
+            $table->dropColumn('deposit_id');
         });
     }
 };
